Java doc ExpenseCategory, proporcion y Nro cuota

+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/DtoExpenseDistributionQuery.java
@@ -16,5 +16,5 @@ public class DtoExpenseDistributionQuery {
     private String ownerFullName;
     private BigDecimal amount;
 
-    private double proportion; //Proporcion
+    private BigDecimal proportion; //Proporcion
 }
